+++ examples/qvm/multishot.c
@@ -10,11 +10,21 @@ void die(char *msg) {
   exit(1);
 }
 
+// We specify a base qubit index of 100 in order to ensure
+// that QVM compresses the number of qubits used in the program.
+// If QVM did not do so, it would fail when trying to allocate
+// memory for 100 qubits. In the examples below, QVM only 
+// really needs to allocate enough memory for 3 qubits.
+const int q0 = 100;
+
 void multishot_with_explicit_ro_indices() {
   quil_program program;
 
-  char *source = "DECLARE ro BIT[3]; X 0; I 1; X 2; MEASURE 0 ro[0]; MEASURE 1 "
-                 "ro[1]; MEASURE 2 ro[2]";
+  char source[200];
+  sprintf(source,
+          "DECLARE ro BIT[3]; X %d; I %d; X %d; MEASURE %d ro[0]; MEASURE %d "
+          "ro[1]; MEASURE %d ro[2]",
+          q0, q0 + 1, q0 + 2, q0, q0 + 1, q0 + 2);
 
   if (quilc_parse_quil(source, &program) != LIBQUIL_ERROR_SUCCESS) {
     LIBQUIL_ERROR("failed to parse quil");
@@ -36,8 +46,8 @@ void multishot_with_explicit_ro_indices() {
 
   qvm_multishot_result qvm_res;
   int num_trials = 10;
-  if (qvm_multishot(program, addresses, num_trials, NULL, NULL, NULL, &qvm_res) !=
-      LIBQUIL_ERROR_SUCCESS) {
+  if (qvm_multishot(program, addresses, num_trials, NULL, NULL, NULL,
+                    &qvm_res) != LIBQUIL_ERROR_SUCCESS) {
     LIBQUIL_ERROR("failed to call qvm_multishot");
     exit(1);
   }
@@ -60,8 +70,11 @@ void multishot_with_explicit_ro_indices() {
 void multishot_with_implicit_ro_indices() {
   quil_program program;
 
-  char *source = "DECLARE ro BIT[3]; X 0; I 1; X 2; MEASURE 0 ro[0]; MEASURE 1 "
-                 "ro[1]; MEASURE 2 ro[2]";
+  char source[200];
+  sprintf(source,
+          "DECLARE ro BIT[3]; X %d; I %d; X %d; MEASURE %d ro[0]; MEASURE %d "
+          "ro[1]; MEASURE %d ro[2]",
+          q0, q0 + 1, q0 + 2, q0, q0 + 1, q0 + 2);
 
   if (quilc_parse_quil(source, &program) != LIBQUIL_ERROR_SUCCESS) {
     LIBQUIL_ERROR("failed to parse quil");
@@ -83,8 +96,8 @@ void multishot_with_implicit_ro_indices() {
   qvm_multishot_result qvm_res;
   int num_trials = 10;
   double gate_noise[] = {0.0, 0.0, 0.0};
-  if (qvm_multishot(program, addresses, num_trials, NULL, NULL, NULL, &qvm_res) !=
-      LIBQUIL_ERROR_SUCCESS) {
+  if (qvm_multishot(program, addresses, num_trials, NULL, NULL, NULL,
+                    &qvm_res) != LIBQUIL_ERROR_SUCCESS) {
     LIBQUIL_ERROR("failed to call qvm_multishot");
     exit(1);
   }
@@ -112,8 +125,11 @@ void multishot_with_implicit_ro_indices() {
 void multishot_with_noise() {
   quil_program program;
 
-  char *source = "DECLARE ro BIT[3]; X 0; I 1; X 2; MEASURE 0 ro[0]; MEASURE 1 "
-                 "ro[1]; MEASURE 2 ro[2]";
+  char source[200];
+  sprintf(source,
+          "DECLARE ro BIT[3]; X %d; I %d; X %d; MEASURE %d ro[0]; MEASURE %d "
+          "ro[1]; MEASURE %d ro[2]",
+          q0, q0 + 1, q0 + 2, q0, q0 + 1, q0 + 2);
 
   if (quilc_parse_quil(source, &program) != LIBQUIL_ERROR_SUCCESS) {
     LIBQUIL_ERROR("failed to parse quil");
@@ -138,7 +154,8 @@ void multishot_with_noise() {
   double gate_noise[] = {0.1, 0.1, 0.1};
   double measurement_noise[] = {0.1, 0.0, 0.0};
   if (qvm_multishot(program, addresses, num_trials, gate_noise,
-                    measurement_noise, NULL, &qvm_res) != LIBQUIL_ERROR_SUCCESS) {
+                    measurement_noise, NULL,
+                    &qvm_res) != LIBQUIL_ERROR_SUCCESS) {
     LIBQUIL_ERROR("failed to call qvm_multishot");
     exit(1);
   }